Error checking works! Figured out the issue was.
The issue was a simple mistake of logic location. I didn't place the error messages and colors in the right if else statement. It was always hitting the else in the if statement where the string numbers are to be converted into integers. It should have been the else in the if statement where it does a string analysis confirming if they have anything else besides improper formatting and decimals.

Learned about behaviors, triggers, and dependency services since I was trying to find the basic way to handle errors. With the limited amount of time and knowldge in regards in using Xamarin, I chose a more efficient route to handle errors.

Primairly learned from this one entry alone on how to handle color changes:

https://forums.xamarin.com/discussion/37751/how-to-change-label-textcolor-with-binding-property-of-viewmodel-in-mvvm

Learned the Regex for integer values from this website:

https://stackoverflow.com/questions/9043551/regex-match-integer-only

When I was creating the formulas to guard the auto generating of the random number list, it converted all the input values into integers, including numbers with decimals. So I stuck with strings since it mirrors what is typed in from the Entry boxes.

A bit more fine tuning on the program. Almost there.