Compiling and archiving the Android apk
Manage to figure out how to create the apk file for android deployment. The references is as follows:

https://channel9.msdn.com/Shows/XamarinShow/Snack-Pack-5-Android-Archive-Manager-for-Visual-Studio
Realizing that I can compile for multiple CPUs
https://docs.microsoft.com/en-us/xamarin/android/deploy-test/release-prep/?tabs=windows#shrink_apk
Had to restart my computer because for some odd reason the project still thinks I was in debug mode. Manage to archive the apk after.

To note, when I tried to install the apk onto my phone, it did not work. However, when I built it directly to my phone, it works perfectly fine.