Adding the ability for a user to input the size of the list
Implemented a third area in the app where the user can input a positive integer to tell the app how many integers it should generate. There are a couple secuirty risks like buffer overflow of the app if the user types in a enormous number but for now we can use it to test performance and see where it breaks.

Another thing to note is that when implementing this feature into the app, there is a issue I noticed with the continuation flow of the code. This happened during the implementation process of the error handling for this new feature.

Typically, code is a continuation flow of instructions, once a function is called, it'll execute the code until there is no more code to execute in the instructions. Now this a gross simplification, there are execeptions, like code interrupts, but usually when the currently executed function calls another function to do a certain task, that function is usually does its job and boomerangs back to the currently executed function and continues executing the rest of the code. No so when implementing the error checking for the new feature.

When I perform a check for the int value tempNumRoll whether it is less than 1 in the ICommand CreateNumListCommand, it'll execute the DisplayErrors function if that number is less than 1. What I wasn't expecting for it though was it not executing the rest of the code after the DisplayErrors function.

What I did believe happen was the following:

-The error code was assigning new string values to BGColor* and ErrorMessage.
-This caused the functions for each of these variables to execute  OnPropertyChanged() function.
-OnPropertyChanged() terminates the currently executed function, since there seems to be termination code.

Just a small thing to note that kinda bugs a little.